// src/service/models/ScriptingParameter.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ScriptingParameter = {
    dataType?: ScriptingParameter.dataType;
    index?: number;
    name?: string;
};
export namespace ScriptingParameter {
    export enum dataType {
        CHARACTER = 'CHARACTER',
        DATETIME = 'DATETIME',
        FLOAT = 'FLOAT',
        INTEGER = 'INTEGER',
    }
}